+++ app/components/qr/image-preview.tsx
@@ -5,64 +5,79 @@ import {
   useState,
   forwardRef,
 } from "react";
-
-export interface ImagePreviewProps {
-  qr: string;
-  logo: string;
-  size: "cable" | "small" | "medium" | "large";
-}
-
-export interface ImagePreviewState {
-  fontSize: number;
-  canvasSize: number;
-  logoSize: number;
-  qrImg: HTMLImageElement | null;
-  logoImg: HTMLImageElement | null;
-}
-
-export interface ImagePreviewRef {
-  exportToPNG(): string;
-}
+import { loadImage } from "~/utils/load-image";
+import type {
+  ImagePreviewProps,
+  ImagePreviewState,
+  ImagePreviewRef,
+} from "./types";
 
 // NOTE: Do not trim the space at the end.
 // It used for spacing between the logo and the text.
-const PROPERTY_OF = "Property of  ";
+const POWERED_BY = "Powered by  ";
 const FONT_SIZE_MAP = {
   cable: 0,
-  small: 12,
-  medium: 15,
-  large: 18,
+  small: 8,
+  medium: 12,
+  large: 16,
 };
 const LOGO_SIZE_MAP = {
   cable: 0,
-  small: 24,
-  medium: 28,
-  large: 32,
+  small: 20,
+  medium: 24,
+  large: 28,
 };
 
-const loadImage = (src: string): Promise<HTMLImageElement | null> =>
-  new Promise((resolve, reject) => {
-    if (src) {
-      const img = new Image();
-      img.src = src;
-      img.onload = () => {
-        resolve(img);
-      };
-      img.onerror = reject;
-    } else {
-      resolve(null);
-    }
-  });
+const drawQr = (ctx: CanvasRenderingContext2D, state: ImagePreviewState) => {
+  const negativeMargin = -10;
+  const { logoSize, fontSize, qrImg, logoImg } = state;
+
+  if (!qrImg) return;
+
+  ctx.fillStyle = "#fff";
+  ctx.setTransform(1, 0, 0, 1, 0, 0);
+  ctx.fillRect(0, 0, ctx.canvas.width, ctx.canvas.height);
+
+  if (state.canvasSize === qrImg.naturalHeight || !logoImg) {
+    ctx.drawImage(qrImg, 0, 0);
+    return;
+  }
+
+  const ctxHalfWidth = ctx.canvas.width / 2;
+  ctx.drawImage(qrImg, ctxHalfWidth - qrImg.naturalWidth / 2, 0);
+
+  ctx.fillStyle = "#9ba5b5";
+  ctx.font = `${fontSize}px Inter`;
+  let metrics = ctx.measureText(POWERED_BY);
+  let textHeight =
+    metrics.actualBoundingBoxAscent + metrics.actualBoundingBoxDescent;
+  let textWidth = metrics.width;
+  let halfTextWidth = textWidth / 2;
+  let textX = ctxHalfWidth - halfTextWidth - logoSize / 2;
+  let textY = qrImg.naturalHeight + logoSize / 2;
+  textY += logoSize / 2 - textHeight / 2;
+  textY += negativeMargin;
+  ctx.fillText(POWERED_BY, textX, textY);
+
+  ctx.drawImage(
+    logoImg,
+    textX + textWidth,
+    qrImg.naturalHeight + negativeMargin,
+    logoSize,
+    logoSize
+  );
+};
 
 export const ImagePreview = forwardRef<ImagePreviewRef, ImagePreviewProps>(
   function ImagePreview({ qr, size, logo }: ImagePreviewProps, ref) {
     const canvasRef = useRef<HTMLCanvasElement>(null);
-    const [state, setState] = useState({
-      fontSize: 15,
+    const [state, setState] = useState<ImagePreviewState>({
+      logoSize: LOGO_SIZE_MAP[size],
+      fontSize: FONT_SIZE_MAP[size],
       canvasSize: 190,
       qrImg: null,
       logoImg: null,
-    } as ImagePreviewState);
+    });
 
     useImperativeHandle(
       ref,
@@ -100,45 +115,9 @@ export const ImagePreview = forwardRef<ImagePreviewRef, ImagePreviewProps>(
     }, [qr, size, logo]);
 
     useEffect(() => {
-      const canvas = canvasRef.current;
-      const ctx = canvas?.getContext("2d");
-      const negativeMargin = -10;
-      const { logoSize, fontSize, qrImg, logoImg } = state;
+      const ctx = canvasRef.current?.getContext("2d");
 
-      if (!ctx || !qrImg) return;
-
-      ctx.fillStyle = "#fff";
-      ctx.setTransform(1, 0, 0, 1, 0, 0);
-      ctx.fillRect(0, 0, ctx.canvas.width, ctx.canvas.height);
-
-      if (state.canvasSize === qrImg.naturalHeight || !logoImg) {
-        ctx.drawImage(qrImg, 0, 0);
-        return;
-      }
-
-      const ctxHalfWidth = ctx.canvas.width / 2;
-      ctx.drawImage(qrImg, ctxHalfWidth - qrImg.naturalWidth / 2, 0);
-
-      ctx.fillStyle = "#9ba5b5";
-      ctx.font = `bold ${fontSize}px Inter`;
-      let metrics = ctx.measureText(PROPERTY_OF);
-      let textHeight =
-        metrics.actualBoundingBoxAscent + metrics.actualBoundingBoxDescent;
-      let textWidth = metrics.width;
-      let halfTextWidth = textWidth / 2;
-      let textX = ctxHalfWidth - halfTextWidth - logoSize / 2;
-      let textY = qrImg.naturalHeight + logoSize / 2;
-      textY += logoSize / 2 - textHeight / 2;
-      textY += negativeMargin;
-      ctx.fillText(PROPERTY_OF, textX, textY);
-
-      ctx.drawImage(
-        logoImg,
-        textX + textWidth,
-        qrImg.naturalHeight + negativeMargin,
-        logoSize,
-        logoSize
-      );
+      if (ctx) drawQr(ctx, state);
     }, [state]);
 
     return (
