+++ app/utils/load-image.ts
@@ -1,0 +1,13 @@
+export const loadImage = (src: string): Promise<HTMLImageElement | null> =>
+  new Promise((resolve, reject) => {
+    if (src) {
+      const img = new Image();
+      img.src = src;
+      img.onload = () => {
+        resolve(img);
+      };
+      img.onerror = reject;
+    } else {
+      resolve(null);
+    }
+  });
